# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clothes E-Commerce API",
        "description": "Comprehensive API Documentation for the full e-commerce system (Customer & Admin routes).",
        "contact": {
            "name": "Figo Team",
            "email": "admin@clothes.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/admin/api/products": {
            "get": {
                "tags": [
                    "Admin - Product"
                ],
                "summary": "Get list of Products",
                "operationId": "47b8980a4cb1ba98159cd71ff57bca38",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin - Product"
                ],
                "summary": "Create/Process Product (store)",
                "operationId": "20c0fe937ed7549b582c75f935a14373",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "base_price",
                                    "stock_quantity",
                                    "subcategory_id",
                                    "slug",
                                    "sku",
                                    "short_description",
                                    "long_description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "base_price": {
                                        "type": "integer"
                                    },
                                    "stock_quantity": {
                                        "type": "integer"
                                    },
                                    "subcategory_id": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "sku": {
                                        "type": "string"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "long_description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/admin/api/products/{id}": {
            "put": {
                "tags": [
                    "Admin - Product"
                ],
                "summary": "Update Product (update)",
                "operationId": "beea5784660f4e70beb94aa258c8001b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "base_price",
                                    "stock_quantity",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "base_price": {
                                        "type": "integer"
                                    },
                                    "stock_quantity": {
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin - Product"
                ],
                "summary": "Delete Product",
                "operationId": "32dd3d65c953f076a6ed4af43b9dd195",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/admin/api/returns": {
            "get": {
                "tags": [
                    "Admin - Return"
                ],
                "summary": "Get list of Returns",
                "operationId": "4089b038c1b1dd2b65c72dfdda212cac",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/admin/api/returns/{id}/status": {
            "put": {
                "tags": [
                    "Admin - Return"
                ],
                "summary": "Update Return (updateStatus)",
                "operationId": "a65c86cf1962475586fc8838452c7c2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/admin/api/wallet/credit": {
            "post": {
                "tags": [
                    "Admin - Wallet"
                ],
                "summary": "addCredit operation",
                "operationId": "5e9211920c69bd67d14790b127237976",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount",
                                    "description"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format **Bearer <token>**",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}